Conflicts:
	deployment/ansible/inventories/group_vars/prod/prod-vault.yml
	deployment/ansible/inventories/group_vars/qa/qa-vault.yml
	deployment/ansible/inventories/group_vars/staging/staging-vault.yml

+++ appserver/neo4japp/blueprints/files.py
@@ -356,6 +356,80 @@ def get_file_info(id: str, project_name: str):
     })
 
 
+@newbp.route('/<string:project_name>/files/<string:file_id>/associated-maps', methods=['GET'])
+@auth.login_required
+@requires_project_permission(AccessActionType.READ)
+def get_associated_maps(file_id: str, project_name: str):
+
+    user = g.current_user
+
+    projects = Projects.query.filter(Projects.project_name == project_name).one_or_none()
+    if projects is None:
+        raise RecordNotFoundException(f'Project {project_name} not found')
+
+    yield user, projects
+
+    query = f"""
+    SELECT
+        DISTINCT
+        p.id
+        , p.hash_id
+        , p.label
+        , p.author
+        , p.dir_id
+    FROM (
+        SELECT
+            p.id
+            , data
+        FROM project p
+        CROSS JOIN json_to_recordset(json_extract_path(graph, 'nodes')) AS data(data JSON)
+        UNION ALL
+        SELECT
+            p.id
+            , data
+        FROM project p
+        CROSS JOIN json_to_recordset(json_extract_path(graph, 'edges')) AS data(data JSON)
+    ) data
+    CROSS JOIN json_to_recordset(json_extract_path(data.data, 'sources')) AS source(url VARCHAR)
+    INNER JOIN project p ON p.id = data.id
+    WHERE
+        url ~ :url_1
+        OR url ~ :url_2
+    """
+
+    results = db.session.execute(
+        query,
+        {
+            'url_1': f'/projects/{project_name}/files/{file_id}(?:#.*)?',
+            'url_2': f'/dt/pdf/{file_id}(?:#.*)?'
+        }
+    ).fetchall()
+
+    directory_project_query_result = db.session.query(
+        Directory.id,
+        Projects.project_name
+    ).filter(
+        Directory.id.in_([row[4] for row in results])
+    ).join(
+        Projects,
+        Projects.id == Directory.projects_id
+    ).all()
+
+    dir_project_map = {
+        dir_id: project_name
+        for (dir_id, project_name) in directory_project_query_result
+    }
+
+    yield jsonify([
+        {
+            'hash_id': row[1],
+            'label': row[2],
+            'author': row[3],
+            'project_name': dir_project_map[row[4]]
+        } for row in results
+    ])
+
+
 @newbp.route('/<string:project_name>/files/<string:id>', methods=['GET', 'PATCH'])
 @auth.login_required
 @requires_project_permission(AccessActionType.READ)
@@ -387,17 +461,24 @@ def get_pdf(id: str, project_name: str):
             if description != file.description:
                 update['description'] = description
 
+            if update:
+                try:
+                    db.session.query(Files).filter(Files.file_id == id).update(update)
+                except SQLAlchemyError:
+                    db.session.rollback()
+                    raise DatabaseError('Failed to update PDF filename and/or description.')  # noqa
+
             curr_fallback = FallbackOrganism.query.get(file.fallback_organism_id)
 
             if not fallback_organism:
-                # fallback organism was removed
-                try:
-                    file.fallback_organism = None
-                    db.session.delete(curr_fallback)
-                    db.session.commit()
-                except SQLAlchemyError:
-                    db.session.rollback()
-                    raise DatabaseError(f'Failed to delete fallback organism from file <{file.file_id}>.')  # noqa
+                if curr_fallback:
+                    # fallback organism was removed
+                    try:
+                        file.fallback_organism = None
+                        db.session.delete(curr_fallback)
+                    except SQLAlchemyError:
+                        db.session.rollback()
+                        raise DatabaseError('Failed to delete fallback organism from the PDF.')  # noqa
             else:
                 if (not curr_fallback or
                     (curr_fallback.organism_name != fallback_organism['organism_name']
@@ -417,14 +498,15 @@ def get_pdf(id: str, project_name: str):
                         file.fallback_organism = new_fallback
                         if curr_fallback:
                             db.session.delete(curr_fallback)
-                        db.session.commit()
                     except SQLAlchemyError:
                         db.session.rollback()
-                        raise DatabaseError(f'There was a problem updating fallback organism from file <{file.file_id}>.')  # noqa
+                        raise DatabaseError('There was a problem updating fallback organism for the PDF.')  # noqa
 
-            if update:
-                db.session.query(Files).filter(Files.file_id == id).update(update)
+            try:
                 db.session.commit()
+            except SQLAlchemyError:
+                db.session.rollback()
+                raise DatabaseError('Unexpected error occurred updating PDF.')
         yield ''
 
     try:
